requirements.txt
Flask==2.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
Flask-WTF==1.0.1
email-validator==1.2.1
python-dateutil==2.8.2